chore: Divide Animation & Style File
